# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Platform API",
        "description": "API complète pour la plateforme d'emploi - Authentification, gestion des offres et candidatures",
        "contact": {
            "email": "support@jobplatform.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/api/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Lister les candidatures",
                "description": "Retourne les candidatures selon le rôle de l'utilisateur",
                "operationId": "5bdb0276efb6e85e52a842b69a9c6601",
                "responses": {
                    "200": {
                        "description": "Liste des candidatures",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Rôle non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{id}/apply": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Postuler à une offre d'emploi",
                "description": "Soumet une candidature pour une offre d'emploi (candidats seulement)",
                "operationId": "7f85f4553a35877cf04b57509bfb063b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données de la candidature",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidature soumise",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationSuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Offre non disponible"
                    },
                    "403": {
                        "description": "Rôle candidat requis"
                    },
                    "409": {
                        "description": "Déjà postulé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applications/{id}": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Voir une candidature spécifique",
                "description": "Affiche les détails d'une candidature spécifique",
                "operationId": "063830ab6918716412802601f8384a05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la candidature",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Applications"
                ],
                "summary": "Mettre à jour le statut d'une candidature",
                "description": "Met à jour le statut d'une candidature (employeur ou admin seulement)",
                "operationId": "0794c698e89bca15fb63f540db6b9d98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nouveau statut",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicationUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUpdateResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Applications"
                ],
                "summary": "Supprimer une candidature",
                "description": "Supprime une candidature (candidat, employeur ou admin)",
                "operationId": "528c6aed37725fba247b8441d8f3f9b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidature supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Application": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cover_letter": {
                        "type": "string",
                        "example": "Je suis très intéressé par ce poste..."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "accepted",
                            "rejected"
                        ],
                        "example": "pending"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "job_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "candidate": {
                        "$ref": "#/components/schemas/User"
                    },
                    "job": {
                        "$ref": "#/components/schemas/Job"
                    }
                },
                "type": "object"
            },
            "ApplicationRequest": {
                "properties": {
                    "cover_letter": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Lettre de motivation...",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApplicationUpdateRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "accepted",
                            "rejected"
                        ],
                        "example": "accepted"
                    }
                },
                "type": "object"
            },
            "ApplicationResponse": {
                "properties": {
                    "role": {
                        "type": "string",
                        "example": "candidate"
                    },
                    "applications": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Application"
                        }
                    }
                },
                "type": "object"
            },
            "ApplicationSuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Application submitted successfully"
                    },
                    "application": {
                        "$ref": "#/components/schemas/Application"
                    }
                },
                "type": "object"
            },
            "ApplicationUpdateResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Application status updated successfully"
                    },
                    "application": {
                        "$ref": "#/components/schemas/Application"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@test.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "Job": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Développeur Full-Stack"
                    },
                    "description": {
                        "type": "string",
                        "example": "Nous recherchons un développeur full-stack expérimenté..."
                    },
                    "company": {
                        "type": "string",
                        "example": "Tech Solutions"
                    },
                    "location": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "employer": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "DeleteResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Job deleted successfully"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}